+++ PostgresSqlAccess/Models/InstructionDataModel.cs
@@ -12,7 +12,7 @@ namespace DataAccess.Models
 	{
 		[Key]
 		public int Id { get; set; }
-		public int? Oz { get; set; }
+		public float? Oz { get; set; }
 		public string? Special { get; set; }
 		public string? DropShopOptions { get; set; }
