+++ PostgresSqlAccess/Models/Interfaces/IGlassDataModel.cs
@@ -4,7 +4,7 @@
     {
         int Id { get; set; }
         string Name { get; set; }
-        int? Oz { get; set; }
+        float? Oz { get; set; }
 
         Uri? Image { get; set; }
     }
