+++ PostgresSqlAccess/Models/Interfaces/IInstructionDataModel.cs
@@ -3,7 +3,7 @@
     public interface IInstructionDataModel
     {
         int Id { get; set; }
-        int? Oz { get; set; }
+        float? Oz { get; set; }
         string? Special { get; set; }
 		public string? DropShopOptions { get; set; }
 	}
